feat(py): refactor UI components, move ProgressManager to a new 'ui' module
