+++ mosaico-sdk-py/src/mosaicolabs/ros_bridge/sequence_extractor.py
@@ -29,9 +29,10 @@ from rosbags.typesys.store import Typestore
 
 from mosaicolabs import Message, MosaicoClient
 from mosaicolabs.logging_config import get_logger, setup_sdk_logging
-from mosaicolabs.ros_bridge.loader import MosaicoLoader, ProgressManager
+from mosaicolabs.ros_bridge.loader import MosaicoLoader
 from mosaicolabs.ros_bridge.qos import get_qos_for_topic
 from mosaicolabs.ros_bridge.registry import ROSTypeRegistry
+from mosaicolabs.ros_bridge.ui import ProgressManager
 
 # Set the hierarchical logger
 logger = get_logger(__name__)
